almost final

+++ sales/api/sales_rest/models.py
@@ -15,7 +15,7 @@ class AutomobileVO(models.Model):
 class Customer(models.Model):
     name = models.CharField(max_length=100)
     address = models.CharField(max_length=100)
-    phone_number = models.PositiveIntegerField(null=True)
+    phone_number = models.BigIntegerField(null=True)
 
     def __str__(self):
         return self.name
@@ -34,7 +34,6 @@ class Salesperson(models.Model):
 class Sale(models.Model):
     sale_price = models.PositiveIntegerField(null=True)
 
-    # FOREIGN KEYS -------------------------------------------------------------
     vin = models.ForeignKey(
         AutomobileVO,
         related_name="sale",
